+++ digitalocean/resource_digitalocean_database_cluster_test.go
@@ -39,6 +39,8 @@ func TestAccDigitalOceanDatabaseCluster_Basic(t *testing.T) {
 						"digitalocean_database_cluster.foobar", "password"),
 					resource.TestCheckResourceAttrSet(
 						"digitalocean_database_cluster.foobar", "urn"),
+					resource.TestCheckResourceAttr(
+						"digitalocean_database_cluster.foobar", "tags.#", "1"),
 				),
 			},
 		},
@@ -233,6 +235,7 @@ resource "digitalocean_database_cluster" "foobar" {
 	size       = "db-s-1vcpu-1gb"
 	region     = "nyc1"
     node_count = 1
+	tags       = ["production"]
 }`
 
 const testAccCheckDigitalOceanDatabaseClusterConfigWithUpdate = `
@@ -243,6 +246,7 @@ resource "digitalocean_database_cluster" "foobar" {
 	size       = "db-s-1vcpu-2gb"
 	region     = "nyc1"
     node_count = 1
+	tags       = ["production"]
 }`
 
 const testAccCheckDigitalOceanDatabaseClusterConfigWithMigration = `
@@ -253,6 +257,7 @@ resource "digitalocean_database_cluster" "foobar" {
 	size       = "db-s-1vcpu-1gb"
 	region     = "lon1"
     node_count = 1
+	tags       = ["production"]
 }`
 
 const testAccCheckDigitalOceanDatabaseClusterConfigWithMaintWindow = `
@@ -263,6 +268,7 @@ resource "digitalocean_database_cluster" "foobar" {
 	size       = "db-s-1vcpu-1gb"
 	region     = "nyc1"
 	node_count = 1
+	tags       = ["production"]
 
 	maintenance_window {
         day  = "friday"
@@ -277,4 +283,5 @@ resource "digitalocean_database_cluster" "foobar" {
 	size       = "db-s-1vcpu-1gb"
 	region     = "nyc1"
     node_count = 1
+	tags       = ["production"]
 }`